Refactor: shared directory separated into several
