+++ components/Adds/Notice/NoticeViewPage.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import useGet from '../../../hooks/useGet';
-import View from '../../Shared/View';
+import View from '../../Element/Shared/View';
 import { Paths } from '../../../constants/paths';
 import { ResponseNotice } from '../../../@types/api/notice';
 import { getNoticeById } from '../../../api/notice';
