+++ components/Admin/Resource/ResourceViewPage.tsx
@@ -2,10 +2,10 @@ import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import { ResponseResource } from '../../../@types/api/resource';
 import useGet from '../../../hooks/useGet';
-import View from '../../Shared/View';
+import View from '../../Element/Shared/View';
 import { getResourceById } from '../../../api/resource';
 import { Paths } from '../../../constants/paths';
-import AdminButton from '../../Shared/AdminButton';
+import AdminButton from '../../Element/Admin/AdminButton';
 import { useRouter } from 'next/router';
 
 export default function ResourceViewPage({ id }: ViewPageProps) {
