+++ components/Element/Adds/Sidebar.tsx
@@ -1,11 +1,11 @@
 import styled from 'styled-components';
-import { Seo } from '../../constants/seo';
-import { Colors } from '../../styles/colors';
-import { Fonts } from '../../styles/fonts';
-import { Paths } from '../../constants/paths';
-import LinkItem from './LinkItem';
+import { Seo } from '../../../constants/seo';
+import { Colors } from '../../../styles/colors';
+import { Fonts } from '../../../styles/fonts';
+import { Paths } from '../../../constants/paths';
+import LinkItem from '../Shared/LinkItem';
 import { useRouter } from 'next/router';
-import { SidebarProps } from '../../@types/shared';
+import { SidebarProps } from '../../../@types/shared';
 
 export default function Sidebar({ path }: SidebarProps) {
 	const router = useRouter();
