+++ components/Element/Shared/Input.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { InputProps } from '../../@types/shared';
-import { Colors } from '../../styles/colors';
+import { InputProps } from '../../../@types/shared';
+import { Colors } from '../../../styles/colors';
 
 export default function Input({ label, value, type, inputRef, onChange }: InputProps) {
 	return (
