+++ components/Layout/AddsLayout.tsx
@@ -1,4 +1,4 @@
-import Sidebar from '../Shared/Sidebar';
+import Sidebar from '../Element/Adds/Sidebar';
 import styled from 'styled-components';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
